dm: core: Create ofnode stubs when OF_REAL is disabled
When there is no device tree there is not point in trying to find nodes,
etc. since they will all be null.

Add static inlines to skip the code in that case.

Unfortunately this makes the file a little convoluted and there are
two inlines for ofnode_is_enabled() and ofnode_first/next_subnode(). But
it seems better than the alternative.

We could also consider splitting up the header file.

Also add a rule in drivers/Makefile to compile ofnode.o when OF_REAL is
enabled but DM is not (for kontron-sl-mx6ul) and move the
ofnode_for_each_compatible_node/prop() macros outside the OF_REAL
condition, since they only use functions that have stubs.

Series-to: concept
Series-cc: heinrich

Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <simon.glass@canonical.com>