Add a simple malloc() implementation for pre-relocation
If we are to have driver model before relocation we need to support some
way of calling memory allocation routines.

The standard malloc() is pretty complicated:

1. It uses some BSS memory for its state, and BSS is not available before
relocation

2. It supports algorithms for reducing memory fragmentation and improving
performace of free(). Before relocation we could happily just not support
free().

3. It includes about 4KB of code (Thumb 2) and 1KB of data. However since
this has been loaded anyway this is not really a problem.

The simplest way to support pre-relocation malloc() is to reserve an area
of memory and allocate it in increasing blocks as needed. This
implementation does this.

To enable it, you need to define the size of the malloc() pool as described
in the README. It will be located above the pre-relocation stack on
supported architectures.

Note that this implementation is only useful on machines which have some
memory available before dram_init() is called - this includes those that
do no DRAM init (like tegra) and those that do it in SPL (quite a few
boards). Enabling driver model preior to relocation for the rest of the
boards is left for a later exercise.

Changes from original commit:
- Squash in commit 'malloc: Redirect to malloc_simple before relocation'
- Modify dlmalloc/dlfree/dlrealloc/dlmemalign (new 2.8.6 names)
- Add #ifdef __UBOOT__ wrapper around the checks
- Redirect to malloc_simple()/memalign_simple() instead of embedding code
- Add declarations for malloc_simple() and memalign_simple() to malloc.h
- Move global_data.h include and DECLARE_GLOBAL_DATA_PTR to top
- Add proper documentation for the two new functions

Signed-off-by: Simon Glass <sjg@chromium.org>
Co-developed-by: Claude <noreply@anthropic.com>
Signed-off-by: Simon Glass <simon.glass@canonical.com>
(cherry picked from commit d59476b6446799c21e64147d86483140154c1886)