video: Document the quirk in truetype_get_cursor_info()
The current code is written to suggest that the cursor position
(priv->pos_ptr) might be respected. In fact a bug means that it never
is, since index is set to priv->pos_ptr, so the if() condition always
fails.

Further, when editing textline objects, the cursor index is always the
same as priv->pos_ptr anyway. Fix the bug but keep the current
behaviour, adding a comment for future readers.

Signed-off-by: Simon Glass <simon.glass@canonical.com>