test: Add a flag for test which need console output
Some tests cannot run when the console is silent.

An example is an expo test which checks text entry into a textline
object. The console must be enabled so that the characters actually
reach the putc_xy() in console_truetype, since in
scene_textline_send_key(), the lineedit restores the vidconsole state,
outputs the character and then saves the state again. If the character
is never output, then the state won't be updated and the lineedit will
be inconsistent.

Rather than having individual tests handle this manually, add an
explicit flag, in the hope that this quirk does not trip anyone else up.

Put the flag next to the existing UTF_CONSOLE flag, since they are
related.

Signed-off-by: Simon Glass <simon.glass@canonical.com>