expo: Correct rendering of textlines when open
When a textline is open, we must render the edit string and show the
cursor. The easiest way to do this is to draw the string again, then
move the cursor back to the correct place.

There is no need to change the font, since the rendering function
handles this.

Signed-off-by: Simon Glass <simon.glass@canonical.com>