gitlab-ci: Remove LAB_ONLY variable
We don't actually need the LAB_ONLY variable, since we can control which
stages run using the other variables (TEST_SUITES, TEST_PY, WORLD_BUILD).

To run only lab jobs, we can set TEST_PY, WORLD_BUILD and TEST_SUITES to
"0". The only advantage to LAB_ONLY is that it avoids doing this. But we
can create a script to handle it.

So drop this special variable.

Signed-off-by: Simon Glass <simon.glass@canonical.com>